api/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
weaviate-client==4.11.1
mistralai==1.5.0
tenacity==8.0.1
aiofiles==23.2.1
pydantic==2.10.6
pydantic-settings==2.8.1
pyjwt==2.8.0
bcrypt==4.0.1
python-multipart==0.0.9